Add: PostWritePage scss추가, postWrite 컴포넌트 삭제 후 postWritePage에 병합, 중복되는 div -> postWriteDiv 컴포넌트로 이동
