--- src/components/PostWrite/PostWrite.module.scss
@@ -1,96 +1,0 @@
-.outWrapper {
-  .titleWrapper {
-    .title {
-    }
-    .autoSaveNoti {
-    }
-  }
-
-  .subHeading {
-  }
-
-  .contentWrapper {
-    .typeOfBusinessWrapper {
-      .categorySelect {
-      }
-      .detailSelect {
-      }
-    }
-
-    .companyNameWrapper {
-      input {
-      }
-    }
-
-    .companyLogoWrapper {
-      .logoUploadWrapper {
-        .imgWrapper {
-          input {
-          }
-          img {
-          }
-        }
-        .logoSizeNoti {
-        }
-      }
-    }
-
-    .companyIntroduce {
-      input {
-      }
-    }
-
-    .companyUrl {
-      input {
-      }
-    }
-
-    .mainWorkingFild {
-      input {
-      }
-    }
-
-    .detailIntroduce {
-      input {
-      }
-    }
-
-    .membership {
-      input {
-      }
-    }
-
-    .ceoInfo {
-      input {
-      }
-    }
-
-    .companyIntroduceFile {
-      .fileUploadWrapper {
-        input {
-        }
-        .fileSizeNoti {
-        }
-      }
-    }
-
-    .usingBranch {
-      .categorySelect {
-      }
-    }
-
-    .agreeWrapper {
-      input {
-      }
-      .agreeContent {
-      }
-    }
-
-    .btnWrapper {
-      .btn {
-      }
-      .resetBtn {
-      }
-    }
-  }
-}
