+++ src/components/SideBar/SideBar.module.scss
@@ -1,6 +1,6 @@
 .sideBarContainer {
   margin: 0;
-  min-width: 16vw;
+  width: 320px;
   background-color: #5bb684;
   color: #fff;
 }
